# frontend/src/api/openapi.json
{
  "components": {
    "schemas": {
...
      "Cancelled": {
        "properties": {
          "cancelled": {
            "title": "Cancelled",
            "type": "string"
          }
        },
        "required": [
          "cancelled"
        ],
        "title": "Cancelled",
        "type": "object"
      },
...
      "Deleted": {
        "properties": {
          "deleted": {
            "title": "Deleted",
            "type": "string"
          }
        },
        "required": [
          "deleted"
        ],
        "title": "Deleted",
        "type": "object"
      },
...
      "FetchRequest": {
        "properties": {
          "model_id": {
            "title": "Model Id",
            "type": "string"
          }
        },
        "required": [
          "model_id"
        ],
        "title": "FetchRequest",
        "type": "object"
      },
      "FetchState": {
        "additionalProperties": true,
        "properties": {
          "bytes_done": {
            "title": "Bytes Done",
            "type": "integer"
          },
          "bytes_total_est": {
            "title": "Bytes Total Est",
            "type": "integer"
          },
          "elapsed_s": {
            "title": "Elapsed S",
            "type": "number"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "progress": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progress"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "model_id",
          "name",
          "status",
          "bytes_done",
          "bytes_total_est",
          "elapsed_s"
        ],
        "title": "FetchState",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/JobRecord"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "JobList",
        "type": "object"
      },
...
      "JobRecord": {
        "properties": {
          "batch_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Batch Id"
          },
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "engine": {
            "default": "comfyui",
            "title": "Engine",
            "type": "string"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "finished_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "log_tail": {
            "items": {
              "type": "string"
            },
            "title": "Log Tail",
            "type": "array"
          },
          "node_times": {
            "additionalProperties": {
              "type": "number"
            },
            "title": "Node Times",
            "type": "object"
          },
          "progress": {
            "default": 0.0,
            "title": "Progress",
            "type": "number"
          },
          "progress_text": {
            "default": "",
            "title": "Progress Text",
            "type": "string"
          },
          "prompt_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Id"
          },
          "recipe": {
            "additionalProperties": true,
            "title": "Recipe",
            "type": "object"
          },
          "result": {
            "additionalProperties": true,
            "title": "Result",
            "type": "object"
          },
          "resumable": {
            "default": true,
            "title": "Resumable",
            "type": "boolean"
          },
          "retry_count": {
            "default": 0,
            "title": "Retry Count",
            "type": "integer"
          },
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "seed": {
            "default": 0,
            "title": "Seed",
            "type": "integer"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "default": "queued",
            "enum": [
              "staged",
              "queued",
              "running",
              "done",
              "failed",
              "cancelled"
            ],
            "title": "Status",
            "type": "string"
          },
          "variant": {
            "default": "full",
            "title": "Variant",
            "type": "string"
          },
          "vram_estimate_gb": {
            "default": 0.0,
            "title": "Vram Estimate Gb",
            "type": "number"
          },
          "wall_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wall S"
          },
          "warm_group": {
            "default": "",
            "title": "Warm Group",
            "type": "string"
          }
        },
        "required": [
          "schema_version",
          "id",
          "batch_id",
          "kind",
          "recipe",
          "seed",
          "engine",
          "status",
          "progress",
          "progress_text",
          "vram_estimate_gb",
          "warm_group",
          "created_at",
          "started_at",
          "finished_at",
          "wall_s",
          "result",
          "error",
          "log_tail",
          "retry_count",
          "resumable",
          "prompt_id",
          "node_times",
          "variant"
        ],
        "title": "JobRecord",
        "type": "object"
      },
      "JobSubmit": {
        "properties": {
          "recipe": {
            "additionalProperties": true,
            "title": "Recipe",
            "type": "object"
          },
          "stage": {
            "default": false,
            "title": "Stage",
            "type": "boolean"
          }
        },
        "required": [
          "recipe"
        ],
        "title": "JobSubmit",
        "type": "object"
      },
      "JobsSubmitted": {
        "properties": {
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobRecord"
            },
            "title": "Jobs",
            "type": "array"
          }
        },
        "required": [
          "jobs"
        ],
        "title": "JobsSubmitted",
        "type": "object"
      },
...
      "ModelListing": {
        "properties": {
          "fetches": {
            "additionalProperties": {
              "$ref": "#/components/schemas/FetchState"
            },
            "title": "Fetches",
            "type": "object"
          },
          "items": {
            "items": {
              "$ref": "#/components/schemas/ModelEntry"
            },
            "title": "Items",
            "type": "array"
          },
          "models_root": {
            "title": "Models Root",
            "type": "string"
          },
          "scanned_at": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scanned At"
          }
        },
        "required": [
          "items",
          "models_root",
          "scanned_at",
          "fetches"
        ],
        "title": "ModelListing",
        "type": "object"
      },
...
      "ModelVerify": {
        "properties": {
          "matches_ledger": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matches Ledger"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "sha256": {
            "title": "Sha256",
            "type": "string"
          }
        },
        "required": [
          "model_id",
          "sha256",
          "matches_ledger"
        ],
        "title": "ModelVerify",
        "type": "object"
      },
...
      "RecipeBody": {
        "properties": {
          "recipe": {
            "additionalProperties": true,
            "title": "Recipe",
            "type": "object"
          }
        },
        "required": [
          "recipe"
        ],
        "title": "RecipeBody",
        "type": "object"
      },
...
      "Released": {
        "properties": {
          "released": {
            "items": {
              "type": "string"
            },
            "title": "Released",
            "type": "array"
          }
        },
        "required": [
          "released"
        ],
        "title": "Released",
        "type": "object"
      },
...
      "T2IPreview": {
        "additionalProperties": true,
        "description": "`POST /recipes/preview` for a t2i recipe: the effective parameters, the exact serialised prompt, missing weights, ETA.",
        "properties": {
          "base_shift": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Base Shift"
          },
          "cfg": {
            "title": "Cfg",
            "type": "number"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "distilled": {
            "title": "Distilled",
            "type": "boolean"
          },
          "estimate": {
            "$ref": "#/components/schemas/Estimate"
          },
          "guidance": {
            "title": "Guidance",
            "type": "number"
          },
          "height": {
            "title": "Height",
            "type": "integer"
          },
          "loras": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Loras",
            "type": "array"
          },
          "max_refs": {
            "title": "Max Refs",
            "type": "integer"
          },
          "max_shift": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Shift"
          },
          "missing": {
            "items": {
              "$ref": "#/components/schemas/MissingWeight"
            },
            "title": "Missing",
            "type": "array"
          },
          "model_id": {
            "title": "Model Id",
            "type": "string"
          },
          "negative_used": {
            "title": "Negative Used",
            "type": "boolean"
          },
          "prompt_mode": {
            "title": "Prompt Mode",
            "type": "string"
          },
          "refs": {
            "title": "Refs",
            "type": "integer"
          },
          "sampler": {
            "title": "Sampler",
            "type": "string"
          },
          "scheduler": {
            "title": "Scheduler",
            "type": "string"
          },
          "serialized_prompt": {
            "title": "Serialized Prompt",
            "type": "string"
          },
          "steps": {
            "title": "Steps",
            "type": "integer"
          },
          "te_device": {
            "title": "Te Device",
            "type": "string"
          },
          "te_id": {
            "title": "Te Id",
            "type": "string"
          },
          "tile_size": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tile Size"
          },
          "tiled_vae": {
            "title": "Tiled Vae",
            "type": "boolean"
          },
          "token_estimate": {
            "title": "Token Estimate",
            "type": "integer"
          },
          "turbo": {
            "title": "Turbo",
            "type": "boolean"
          },
          "turbo_strength": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turbo Strength"
          },
          "weight_dtype": {
            "title": "Weight Dtype",
            "type": "string"
          },
          "width": {
            "title": "Width",
            "type": "integer"
          },
          "word_count": {
            "title": "Word Count",
            "type": "integer"
          }
        },
        "required": [
          "serialized_prompt",
          "prompt_mode",
          "width",
          "height",
          "steps",
          "guidance",
          "cfg",
          "sampler",
          "scheduler",
          "turbo",
          "distilled",
          "word_count",
          "token_estimate",
          "refs",
          "max_refs",
          "model_id",
          "te_id",
          "negative_used",
          "weight_dtype",
          "te_device",
          "tiled_vae",
          "missing",
          "estimate",
          "count"
        ],
        "title": "T2IPreview",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "loom2 orchestrator",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/jobs": {
      "get": {
        "operationId": "jobs_list_jobs_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Jobs List",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "operationId": "jobs_submit_jobs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/JobSubmit"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobsSubmitted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Jobs Submit",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}": {
      "delete": {
        "operationId": "job_delete_jobs__job_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Deleted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Delete",
        "tags": [
          "jobs"
        ]
      },
      "get": {
        "operationId": "job_get_jobs__job_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRecord"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Get",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/cancel": {
      "post": {
        "operationId": "job_cancel_jobs__job_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cancelled"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Cancel",
        "tags": [
          "jobs"
        ]
      }
    },
    "/jobs/{job_id}/release": {
      "post": {
        "operationId": "job_release_jobs__job_id__release_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Released"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Job Release",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/models": {
      "get": {
        "operationId": "models_models_get",
        "parameters": [
          {
            "in": "query",
            "name": "include_retired",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Retired",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelListing"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Models",
        "tags": [
          "models"
        ]
      }
    },
    "/models/fetch": {
      "post": {
        "operationId": "models_fetch_models_fetch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FetchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FetchState"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Models Fetch",
        "tags": [
          "models"
        ]
      }
    },
...
    "/models/{model_id}/verify": {
      "post": {
        "operationId": "models_verify_models__model_id__verify_post",
        "parameters": [
          {
            "in": "path",
            "name": "model_id",
            "required": true,
            "schema": {
              "title": "Model Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelVerify"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Models Verify",
        "tags": [
          "models"
        ]
      }
    },
...
    "/recipes/preview": {
      "post": {
        "description": "09 §3a: the exact string the engine will receive, effective parameters, ETA and VRAM fit, missing weights.",
        "operationId": "recipe_preview_recipes_preview_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RecipeBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/T2IPreview"
                    },
                    {
                      "$ref": "#/components/schemas/I2VPreview"
                    }
                  ],
                  "title": "Response Recipe Preview Recipes Preview Post"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recipe Preview",
        "tags": [
          "meta"
        ]
      }
    },
...
  }
}